+++ src/SharpFunctional.MSSQL/FunctionalMsSqlDb.cs
@@ -1,13 +1,12 @@
 using System.Data;
 using System.Diagnostics;
-using LanguageExt;
-using LanguageExt.Common;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using SharpFunctional.MsSql.Common;
 using SharpFunctional.MsSql.Dapper;
 using SharpFunctional.MsSql.Ef;
-using static LanguageExt.Prelude;
+using SharpFunctional.MsSql.Functional;
+using static SharpFunctional.MsSql.Functional.Prelude;
 
 namespace SharpFunctional.MsSql;